create a reusable validator for validateNetwork and validateAdminLevels
