+++ src/device-registry/validators/common/index.js
@@ -1,0 +1,7 @@
+const { validateNetwork } = require("./network.validators");
+const { validateAdminLevels } = require("./admin-levels.validators");
+
+module.exports = {
+  validateNetwork,
+  validateAdminLevels,
+};
