+++ src/device-registry/validators/locations.validators.js
@@ -13,19 +13,7 @@ const { HttpError } = require("@utils/shared");
 const httpStatus = require("http-status");
 const createAirQloudUtil = require("@utils/airqloud.util");
 const isEmpty = require("is-empty");
-const NetworkModel = require("@models/Network");
-
-const validNetworks = async () => {
-  const networks = await NetworkModel("airqo").distinct("name");
-  return networks.map((network) => network.toLowerCase());
-};
-
-const validateNetwork = async (value) => {
-  const networks = await validNetworks();
-  if (!networks.includes(value.toLowerCase())) {
-    throw new Error("Invalid network");
-  }
-};
+const { validateNetwork, validateAdminLevels } = require("@validators/common");
 
 const commonValidations = {
   tenant: [
